upload images refactor
Upload to tmp s3 bucket directory and post link the image to the brand
new claim. Post processing the image in order to copy it to the proper
directory and then remove the tmp file.

The versions are always created, tmp versions too. It should have a
check so the versions are only created once at the end of the process.